CB#7 - Refactoring - Remove users feature logic from Infrastructure project
