--- src/crypto-bank.Infrastructure/Features/Users/Exceptions/UserAlreadyExistsException.cs
@@ -1,11 +1,0 @@
-namespace crypto_bank.Infrastructure.Features.Users.Exceptions;
-
-public class UserAlreadyExistsException : Exception
-{
-    public UserAlreadyExistsException(string email)
-    {
-        Email = email;
-    }
-
-    public string Email { get; }
-}
