--- src/crypto-bank.Infrastructure/Features/Users/Exceptions/UserNotFoundException.cs
@@ -1,11 +1,0 @@
-namespace crypto_bank.Infrastructure.Features.Users.Exceptions;
-
-public class UserNotFoundException : Exception
-{
-    public UserNotFoundException(string email)
-    {
-        Email = email;
-    }
-
-    public string Email { get; }
-}
