+++ src/crypto-bank.Infrastructure/InfrastructureServiceCollectionExtensions.cs
@@ -1,6 +1,5 @@
 using crypto_bank.Infrastructure.Common;
 using crypto_bank.Infrastructure.Features.Auth;
-using crypto_bank.Infrastructure.Features.Users;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -12,8 +11,6 @@ public static class InfrastructureServiceCollectionExtensions
         this IServiceCollection services,
         ConfigurationManager configuration)
     {
-        services.AddUsers(configuration);
-
         services.AddAuth(configuration);
 
         services.AddSingleton<IPasswordHasher, PasswordHasher>();
