+++ src/crypto-bank.WebAPI/Features/Users/DependencyRegistration/ServiceCollectionExtensions.cs
@@ -1,3 +1,4 @@
+using crypto_bank.WebAPI.Features.Users.Options;
 using crypto_bank.WebAPI.Features.Users.Requests;
 using FluentValidation;
 
@@ -5,10 +6,12 @@ namespace crypto_bank.WebAPI.Features.Users.DependencyRegistration;
 
 public static class ServiceCollectionExtensions
 {
-    public static IServiceCollection AddUsers(this IServiceCollection services)
+    public static IServiceCollection AddUsers(this IServiceCollection services, IConfiguration configuration)
     {
         services.AddScoped<IValidator<Register.Request>, Register.RequestValidator>();
         services.AddScoped<IValidator<UpdateRoles.Request>, UpdateRoles.RequestValidator>();
+
+        services.Configure<UsersOptions>(configuration.GetSection("Features:Users"));
 
         return services;
     }
