+++ src/crypto-bank.WebAPI/Features/Users/Errors/RegisterUserLogicConflictError.cs
@@ -1,0 +1,8 @@
+namespace crypto_bank.WebAPI.Features.Users.Errors;
+
+internal static class RegisterUserLogicConflictError
+{
+    private const string Prefix = "register_user_";
+
+    public const string AlreadyExists = Prefix + "already_exists";
+}
